refactor: rename task functions file
